Add messages to otherwise confusing errors
If these have a mismatch, they would simply say "X matches expected Y",
which is not that helpful when there are multiple of the same kind of
thing. To make things a bit faster to debug, there's now a message
included that is outputted added to assertions that may not be that
straightforward.