Consider unarmored movement modifiers
Unarmored movement feats, like "Natural Armor", changes both the AC
value (10 -> 12) and what ability score modifier that should be added to
the score (DEX -> CON). We don't seem to have to care about changing the
value, as it seems like the modifier matches it so we can just add it to
the list. However we have to change the ability score modifier used so
I make sure to change it for Natural Armor. Might be other feats that
wants to the same treatment, but this is the one I've found so far.