Separate character expected character data
Having character data directly in the data provider method felt clunky,
especially the way I had structured it where each value in the array
didn't mention what it was related to. Instead I have decided to move
the expected data out from the data provider function to keep in
separate files. I also added keys to each value to make it easier to
tell what they are for.

This way, to add a new character to be run in tests I can add two
files (api response and expected data) and the tests should
automatically pick the new character up in the data provider.