Assert movement speeds
Since the character data is PHP objects and not just a list of values
like the expected data is (which is in JSON format), I encode both the
expected and actual values to JSON before assertion to make sure that
their formats match. I'm pretty sure there is a better way to handle it,
but it will at least compare the values for now.