Include features in output
Some features are already bound to the class, so we can just grab them
from there, while some are bound to the character directly. There are
some racial features not included here for now, but those will most
likely be done when I focus on adding racial data to the output.