Rename project to dnd-character
Previously I renamed the namespace, however I named it something that
makes it sound like it only generates sheets which is not completely
true. I find it more sensible to name remove sheet from the name as it's
more about getting a readable format for different providers that can be
used for creating character sheets.